mcp_system/requirements.txt
firebase-admin
python-dotenv
mcp
uv
streamlit
openai
nest_asyncio